requirements.txt
fastapi>=0.88.0,<0.89
pymongo>=4.3.2,<4.4
uvicorn>=0.20.0,<0.21
pydantic>=1.10.4,<1.11
beanie>=1.16.6,<1.17
python-dotenv>=0.21.0,<0.22
